revert: use wcag2 color contrast
